requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
google-cloud-storage==2.13.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1